organized tests and added tests for the entries_post_slug view
